Add Configuration
Allow configuration of each Check via a YAML config file. This file is
keyed by the "moniker" of the Check, and can provide options on how it
should be applied.

As an example, allow the WikidataIdentifiers check to vary the column
name it's interested in, and whether capitalisation of the column is
important.