Add a Report class
This class finds all the Checks relevant to a given file (through an
'AppliesTo' configuration setting), and collates the errors from running
each of those checks.

Here I'm assuming that all Checks will live in the `check` directory,
and so can be loaded dynamically, using `require_rel`. There doesn't
seem to be any way to know what was actually loaded by that, so we then
have to find which subclasses of Commons::Integrity::Check::Base exist.