require activesupport
Primarily we want it here for its nice implementation of
Class.descendants, so we can find all the Checks that are loaded.

We _could_ inline that ourselves — it's essentially a one-liner — but we
may as well have someone else maintain that, and activesupport gives us
lots of other things as well.