Report should ignore checks that haven't been configured.
Only checks that are in the integrity Config file should be checked for
relevancy against files.

This means that newly-added check implementations won't cause errors for old
config files, and that when testing, each example config file need not list all
checks.

This refactors relevant_checks to use a new `Config#checks` method, with testing
for applicability of configured tests now done with `Check::Base#applies?`.