Don't conflict with the name of the commons-builder Commons module
The current primary use case of these integrity checks is from
commons-builder, which contains a `module Commons`. This means that if
you require any file containing `class Commons` you will get the error:

  TypeError: Commons is not a class

This can be fixed by making `Commons` a module in this repository, for
consistency with commons-builder.

This commit also changes other namespacing classes that I wouldn't
expect to be instantiated from `class` to `module`.