Check that all area-associated positions have detailed role data
In the future, in the course of building data for a country, we will try
to find from Wikidata information about each position that's associated
with an area (e.g. the associated chamber, the labels of the position, etc.)

However, we might not be able to find all the information we require, in
which case there will be positions mentioned in
boundaries/build/index.json that are not found in
boundaries/build/position-data.json. The check implemented in this
commit produces an error if that's the case.

Fixes #18